+++ adr/1 - Use Radix-36 Strings as Timestamps.md
