+++ adr/2 - An Entry may utilize zero or many Defs.md
