+++ adr/Entry Period Scope.md
@@ -2,6 +2,7 @@
 status: proposed
 date: 2024-09-11
 ---
+#TODO - decide for sure on this.
 
 # All Entries are at the Lowest-Level Scope
