+++ test/test_datasets.ts
@@ -241,36 +241,37 @@ export const journalToOverviewAndIndex: dj.DataJournal = {
     entries: [readEntry, readAndWorkedOutWithName, lifted, deletedEntry]
 }
 
-export const expectedOverview: dj.Overview = {
-    updated: {
-        epochStr: "m0zzzzzz", //from the deleted entry
-        localeStr: "9/12/2024, 7:46:32 PM", //translated
-        isoStr: "2024-09-13T00:46:32.447Z", //translated
-    },
-    counts: {
-        defs: 3,
-        activeEntries: 3,
-        deletedEntries: 1
-    },
-    index: {
-        defMap: {
-            "BOOK_READ_NAME": 0,
-            "WORKOUT_TYPE": 1,
-            "WORKOUT_NAME": 2
-        },
-        entryMap: {
-            "M0OFGFIO_GJLP": 0,
-            "M0OGACOF_3FJK": 1,
-            "M0OGDGGG_CA3T": 2,
-            "M0OFACHO_POAX": 3,
-        },
-        defLblToIdMap: {
-            "Book": "BOOK_READ_NAME",
-            "Workout Name": "WORKOUT_NAME",
-            "Workout Type": "WORKOUT_TYPE",
-        }
-    }
-}
+//#TODO - fix this by importing Overview or whatever if you're still doing that
+// export const expectedOverview: dj.Overview = {
+//     updated: {
+//         epochStr: "m0zzzzzz", //from the deleted entry
+//         localeStr: "9/12/2024, 7:46:32 PM", //translated
+//         isoStr: "2024-09-13T00:46:32.447Z", //translated
+//     },
+//     counts: {
+//         defs: 3,
+//         activeEntries: 3,
+//         deletedEntries: 1
+//     },
+//     index: {
+//         defMap: {
+//             "BOOK_READ_NAME": 0,
+//             "WORKOUT_TYPE": 1,
+//             "WORKOUT_NAME": 2
+//         },
+//         entryMap: {
+//             "M0OFGFIO_GJLP": 0,
+//             "M0OGACOF_3FJK": 1,
+//             "M0OGDGGG_CA3T": 2,
+//             "M0OFACHO_POAX": 3,
+//         },
+//         defLblToIdMap: {
+//             "Book": "BOOK_READ_NAME",
+//             "Workout Name": "WORKOUT_NAME",
+//             "Workout Type": "WORKOUT_TYPE",
+//         }
+//     }
+// }
 //#endregion
 
 //#region --- EXPECTED GROUPING
